Default to static libc++ to match godot-cpp
Didn't run into this earlier because I was forcing it static for all
mingw builds, now both compilers are controlled by a new build
variable. The name of the variable was chosen to match godot-cpp.

As of this commit, building with MSVC appears to be fully working. It
does produce a lot of type mismatch warnings which need to be
investigated, I'm not sure if mingw is actually producing a different
result or if it's just quieter about it.