Add VREvent signal support
This uses a macro to ADD_SIGNAL and track the data types associated with
each EVREventType in a single action, preventing any risk of signals
getting out of sync with handled events as more are added by Valve.

The OpenVREventHandler is only ever intended to be used as the parent
class of the autoload singleton. This allows the macro magic to be used
for generating signal names instead of having to manually define them in
GDScript. There should also only ever be one handler, and the interface
for hooking it to openvr_data reflects this.

Not yet sending along the event data, just an empty placeholder
dictionary. I'm torn on whether it would be worth the effort to create
real classes to go with each event data type, vs. just returning a
dictionary for them all.