Add register_event_signal to allow receiving unsupported events
When new events appear in later releases of OpenVR, they tend to have an
existing data type. This allows users of the extension to register
signals for these new events instead of having to wait for a new release
of godot_openvr to take advantage of them. For example:

  OpenVRInterface.register_event_signal(539, OpenVREventHandler.Overlay, "OverlayCreated")
  OpenVRInterface.connect("OverlayCreated", print)

This could also be used to add signals for events in the vendor-specific
space starting at ID 10000.

I haven't added any handling for trying to register an existing signal,
because it's already handled as an error by godot. Logging an error
seems like the right thing to do anyway, letting the developer know that
their signal either already exists, or was just added in a new release.